requirements.txt
langchain
llama-cpp-python
chromadb
huggingface_hub
pypdf
tiktoken
fastapi
uvicorn
streamlit
langchain
llama-cpp-python
python-dotenv